Add readable

+++ src/remove_worm.sh
@@ -1,1 +1,5 @@
-rm -rf /home/victim/.etc && rm -rf /home/victim/.Launch_Attack && echo "Worm deleted." || echo "Failed to delete the worm."
+if [ -d "/home/victim/.etc" || -d "/home/victim/.Launch_Attack" ]; then 
+    rm -rf /home/victim/.etc && rm -rf /home/victim/.Launch_Attack &&　echo "Worm deleted."　|| echo "Failed to delete the worm."
+else 
+    echo "Worm not found."
+fi
